fix: change score type from float64 to uint64
- MySQL: BIGINT UNSIGNED (부동소수점 오차 제거)
- Redis ZADD: float64(score) 캐스팅 (Redis 내부 규격)
- HSET staging: FormatUint/ParseUint 사용
- scoreRecord: CreatedAt 필드 추가

+++ apps/ranking-service/internal/adapter/in/http/dto.go
@@ -1,9 +1,9 @@
 package http
 
 type RankEntryResponse struct {
-	UserID string  `json:"userId"`
-	Score  float64 `json:"score"`
-	Rank   int64   `json:"rank"`
+	UserID string `json:"userId"`
+	Score  uint64 `json:"score"`
+	Rank   int64  `json:"rank"`
 }
 
 type TopRankingsResponse struct {
