+++ apps/score-service/internal/adapter/in/http/dto.go
@@ -1,11 +1,11 @@
 package http
 
 type SubmitScoreRequest struct {
-	UserID string  `json:"userId" binding:"required"`
-	Score  float64 `json:"score"  binding:"required"`
+	UserID string `json:"userId" binding:"required"`
+	Score  uint64 `json:"score"  binding:"required"`
 }
 
 type SubmitScoreResponse struct {
-	UserID string  `json:"userId"`
-	Score  float64 `json:"score"`
+	UserID string `json:"userId"`
+	Score  uint64 `json:"score"`
 }
